# kardo_swagger.json
{"openapi":"3.0.1","info":{"title":"Kardo Api","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Данные User для авторизованных пользователей","description":"Управление данными User"},{"name":"Подписки авторизованного User","description":"Управление подписками"},{"name":"Данные User для неавторизованных пользователей","description":"Управление данными User"},{"name":"Сообщения пользователей","description":"Управление сообщениями пользователя"}],"paths":{"/users/user/messages/{senderId}/{recipientId}":{"get":{"tags":["Сообщения пользователей"],"summary":"Получение списка сообщений Отправителя и Получателя","operationId":"getAllMessages","parameters":[{"name":"senderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"recipientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageWithShortUser"}}}}}}},"post":{"tags":["Сообщения пользователей"],"summary":"Создание сообщения","operationId":"create","parameters":[{"name":"senderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"recipientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewMessage"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/user/users/subscriptions/{userId}/{followingId}":{"post":{"tags":["Подписки авторизованного User"],"summary":"Создание подписки","operationId":"create_1","parameters":[{"name":"userId","in":"path","description":"User Id","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"followingId","in":"path","description":"Following Id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK"}}},"delete":{"tags":["Подписки авторизованного User"],"summary":"Удаление подписки","operationId":"deleteSubscription","parameters":[{"name":"userId","in":"path","description":"User Id","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"followingId","in":"path","description":"Following Id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK"}}}},"/public/users":{"post":{"tags":["Данные User для неавторизованных пользователей"],"summary":"Создание пользователя","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewUserRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}}}}},"/public/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewUserRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}}}}},"/public/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/users/user/messages/status/{messageId}/{userId}":{"patch":{"tags":["Сообщения пользователей"],"summary":"Изменение статуса сообщения: не прочитано/прочитано","operationId":"updateMessageStatus","parameters":[{"name":"messageId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageWithShortUser"}}}}}}},"/user/users/{userId}":{"get":{"tags":["Данные User для авторизованных пользователей"],"summary":"Получение данных профиля User","operationId":"getUserById","parameters":[{"name":"userId","in":"path","description":"ID Пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}}}},"delete":{"tags":["Данные User для авторизованных пользователей"],"summary":"Удаление профиля User","operationId":"deleteUserById","parameters":[{"name":"userId","in":"path","description":"ID Пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}}}},"patch":{"tags":["Данные User для авторизованных пользователей"],"summary":"Обновление данных пользователя","operationId":"updateUser","parameters":[{"name":"userId","in":"path","description":"ID Пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}}}}},"/public/users/search":{"get":{"tags":["Данные User для неавторизованных пользователей"],"summary":"Поиск User (публичный профиль) по/без параметров","operationId":"findUser","parameters":[{"name":"nickname","in":"query","required":false,"schema":{"type":"string"}},{"name":"firstName","in":"query","required":false,"schema":{"type":"string"}},{"name":"lastName","in":"query","required":false,"schema":{"type":"string"}},{"name":"middleName","in":"query","required":false,"schema":{"type":"string"}},{"name":"birthday","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"country","in":"query","required":false,"schema":{"type":"string"}},{"name":"region","in":"query","required":false,"schema":{"type":"string"}},{"name":"city","in":"query","required":false,"schema":{"type":"string"}},{"name":"style","in":"query","required":false,"schema":{"type":"string"}},{"name":"from","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserShortPage"}}}}}}}},"/public/users/page/{userId}":{"get":{"tags":["Данные User для неавторизованных пользователей"],"summary":"Получение публичного профиля User","operationId":"getUserPageById","parameters":[{"name":"userId","in":"path","description":"ID Пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserPage"}}}}}}},"/public/users/checkNickname":{"get":{"tags":["Данные User для неавторизованных пользователей"],"summary":"Проверка уникальности Nickname","operationId":"checkNickname","parameters":[{"name":"nickname","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/public/users/checkEmail":{"get":{"tags":["Данные User для неавторизованных пользователей"],"summary":"Проверка уникальности email","operationId":"checkEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}}},"components":{"schemas":{"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"}}},"NewMessage":{"required":["text"],"type":"object","properties":{"text":{"type":"string"}}},"Message":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sender":{"$ref":"#/components/schemas/User"},"recipient":{"$ref":"#/components/schemas/User"},"text":{"type":"string"},"dateMessage":{"type":"string","format":"date-time"},"wasRead":{"type":"boolean"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string"},"nickname":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"birthday":{"type":"string","format":"date"},"country":{"type":"string"},"region":{"type":"string"},"city":{"type":"string"},"phone":{"type":"string"},"photoLink":{"type":"string"},"backgroundLink":{"type":"string"},"pageLink":{"type":"string"},"aboutMe":{"type":"string"},"style":{"type":"string","enum":["WORKOUT","HIPHOP","BREAKING","BMX","GRAFFITY","DJING"]},"userRole":{"type":"string","enum":["ADMIN","USER"]},"state":{"type":"string","enum":["ACTIVE","DELETED"]}}},"NewUserRequest":{"required":["birthday","city","country","email","firstName","lastName","middle_name","nickname","password","region"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"nickname":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middle_name":{"type":"string"},"birthday":{"type":"string","format":"date"},"country":{"type":"string"},"region":{"type":"string"},"city":{"type":"string"},"phone":{"type":"string"},"photoLink":{"type":"string"},"backgroundLink":{"type":"string"},"pageLink":{"type":"string"},"style":{"type":"string"},"aboutMe":{"type":"string"}}},"UserProfile":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"nickname":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"birthday":{"type":"string","format":"date"},"country":{"type":"string"},"region":{"type":"string"},"city":{"type":"string"},"phone":{"type":"string"},"photoLink":{"type":"string"},"backgroundLink":{"type":"string"},"pageLink":{"type":"string"},"style":{"type":"string","enum":["WORKOUT","HIPHOP","BREAKING","BMX","GRAFFITY","DJING"]},"aboutMe":{"type":"string"},"state":{"type":"string","enum":["ACTIVE","DELETED"]}}},"AuthRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"MessageWithShortUser":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sender":{"$ref":"#/components/schemas/UserShortPage"},"recipient":{"$ref":"#/components/schemas/UserShortPage"},"text":{"type":"string"},"dateMessage":{"type":"string","format":"date-time"},"wasRead":{"type":"boolean"}}},"UserShortPage":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"photoLink":{"type":"string"},"style":{"type":"string","enum":["WORKOUT","HIPHOP","BREAKING","BMX","GRAFFITY","DJING"]},"state":{"type":"string","enum":["ACTIVE","DELETED"]}}},"UserUpdateRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"nickname":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"birthday":{"type":"string","format":"date"},"country":{"type":"string"},"region":{"type":"string"},"city":{"type":"string"},"phone":{"type":"string"},"photoLink":{"type":"string"},"backgroundLink":{"type":"string"},"pageLink":{"type":"string"},"style":{"type":"string"},"aboutMe":{"type":"string"}}},"UserPage":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"birthday":{"type":"string","format":"date"},"country":{"type":"string"},"region":{"type":"string"},"city":{"type":"string"},"photoLink":{"type":"string"},"backgroundLink":{"type":"string"},"pageLink":{"type":"string"},"style":{"type":"string","enum":["WORKOUT","HIPHOP","BREAKING","BMX","GRAFFITY","DJING"]},"aboutMe":{"type":"string"},"state":{"type":"string","enum":["ACTIVE","DELETED"]},"myFollowers":{"type":"array","items":{"$ref":"#/components/schemas/UserShortPage"}},"ifollowings":{"type":"array","items":{"$ref":"#/components/schemas/UserShortPage"}}}}}}}